migrate: move vamcSystem to products directory
- Migrated vamcSystem product files to src/products/vamcSystem/
- Moved and renamed index.tsx -> template.tsx
- Moved and renamed index.test.tsx -> template.test.tsx
- Moved ManageYourHealthLinks component and test files
- Moved StoryTeaser component and test files
- Moved query.ts from src/data/queries/ to src/products/vamcSystem/
- Moved test file and renamed vamcSystem.test.tsx -> query.test.ts
- Moved formatted type from src/types/formatted/ to src/products/vamcSystem/formatted-type.ts
- Moved mock file from src/mocks/ to src/products/vamcSystem/mock.json
- Updated all import statements across the codebase
- Fixed import paths in migrated files
- All TypeScript errors resolved
- All tests passing (7 query tests, 30 template tests, 4 ManageYourHealthLinks tests, 12 StoryTeaser tests)
- Updated migration guide with completed status
- Added CODEOWNERS entry for new product directory
- Verified dev server runs and vamcSystem pages load correctly (Washington DC & Pittsburgh Health Care systems)
