+++ src/data/queries/index.ts
@@ -45,7 +45,7 @@ import * as StoryListing from '@/products/storyListing/query'
 import * as SupportServices from '@/data/queries/supportServices'
 import * as Table from '@/data/queries/table'
 import * as VamcEhr from '@/data/queries/vamcEhr'
-import * as VamcSystem from '@/data/queries/vamcSystem'
+import * as VamcSystem from '@/products/vamcSystem/query'
 import * as VamcSystemVaPolice from '@/products/vamcSystemVaPolice/query'
 import * as VetCenter from '@/products/vetCenter/query'
 import * as VetCenterHealthServices from '@/data/queries/vetCenterHealthServices'
