+++ src/pages/[[...slug]].tsx
@@ -53,7 +53,7 @@ import { StaffProfile as FormattedStaffProfile } from '@/products/staffProfile/f
 import { StoryListing as FormattedStoryListing } from '@/products/storyListing/formatted-type'
 import { VetCenter as FormattedVetCenter } from '@/products/vetCenter/formatted-type'
 import { HealthCareLocalFacility as FormattedHealthCareLocalFacility } from '@/types/formatted/healthCareLocalFacility'
-import { VamcSystem as FormattedVamcSystem } from '@/types/formatted/vamcSystem'
+import { VamcSystem as FormattedVamcSystem } from '@/products/vamcSystem/formatted-type'
 import { VamcSystemVaPolice as FormattedVamcSystemVaPolice } from '@/products/vamcSystemVaPolice/formatted-type'
 import { LeadershipListing as FormattedLeadershipListing } from '@/products/leadershipListing/formatted-type'
 import { VetCenterLocationListing as FormattedVetCenterLocationListing } from '@/products/vetCenterLocationListing/formatted-type'
@@ -77,7 +77,7 @@ import { VetCenter } from '@/products/vetCenter/template'
 import { Wrapper } from '@/templates/layouts/wrapper'
 import { HealthCareLocalFacility } from '@/templates/layouts/healthCareLocalFacility'
 import { DoNotPublishError } from '@/lib/drupal/query'
-import { VamcSystem } from '@/templates/layouts/vamcSystem'
+import { VamcSystem } from '@/products/vamcSystem/template'
 import { VamcSystemVaPolice } from '@/products/vamcSystemVaPolice/template'
 import { LeadershipListing } from '@/products/leadershipListing/template'
 import { VbaFacility } from '@/templates/layouts/vbaFacility'
