+++ src/products/vamcSystem/StoryTeaser.test.tsx
