+++ src/products/vamcSystem/StoryTeaser.tsx
