+++ src/products/vamcSystem/formatted-type.ts
@@ -1,10 +1,10 @@
 import { MediaImage } from '@/types/formatted/media'
 import { Administration } from '@/types/formatted/administration'
 import { SideNavMenu } from '@/types/formatted/sideNav'
-import { PublishedEntity } from './publishedEntity'
-import { HealthCareLocalFacility } from './healthCareLocalFacility'
-import { FormattedRelatedLinks } from './relatedLinks'
-import { NodeHealthCareRegionPage } from '../drupal/node'
+import { PublishedEntity } from '@/types/formatted/publishedEntity'
+import { HealthCareLocalFacility } from '@/types/formatted/healthCareLocalFacility'
+import { FormattedRelatedLinks } from '@/types/formatted/relatedLinks'
+import { NodeHealthCareRegionPage } from '@/types/drupal/node'
 import { LovellChildVariant } from '@/lib/drupal/lovell/types'
 import { NewsStoryTeaser } from '@/products/newsStory/formatted-type'
 import { FacilitySocialLinksProps } from '@/templates/layouts/healthCareLocalFacility/FacilitySocialLinks'
