+++ src/products/vamcSystem/mock.json
