+++ src/products/vamcSystem/template.test.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
-import drupalMockData from '@/mocks/vamcSystem.mock.json'
+import drupalMockData from '@/products/vamcSystem/mock.json'
 import drupalMockFacilityData from '@/mocks/healthCareLocalFacility.mock'
 import drupalMockStoryData from '@/products/newsStory/mock.json'
-import { VamcSystem } from './index'
-import { VamcSystem as FormattedVamcSystem } from '@/types/formatted/vamcSystem'
-import { formatter } from '@/data/queries/vamcSystem'
+import { VamcSystem } from './template'
+import { VamcSystem as FormattedVamcSystem } from '@/products/vamcSystem/formatted-type'
+import { formatter } from '@/products/vamcSystem/query'
 import { DrupalMenuLinkContent } from 'next-drupal'
 import { LOVELL } from '@/lib/drupal/lovell/constants'
 import mockEventData from '@/products/event/mock.json'
