+++ src/products/vamcSystem/template.tsx
@@ -1,4 +1,4 @@
-import { VamcSystem as FormattedVamcSystem } from '@/types/formatted/vamcSystem'
+import { VamcSystem as FormattedVamcSystem } from '@/products/vamcSystem/formatted-type'
 import { MediaImage } from '@/templates/common/mediaImage'
 import { ContentFooter } from '@/templates/common/contentFooter'
 import { LovellStaticPropsResource } from '@/lib/drupal/lovell/types'
@@ -8,10 +8,10 @@ import { FacilityListing } from '@/templates/components/facilityListing'
 import { RelatedLinks } from '@/templates/common/relatedLinks'
 import { RegionalTopTasks } from '@/templates/components/topTasks'
 import { LOVELL } from '@/lib/drupal/lovell/constants'
-import { ManageYourHealthLinks } from '@/templates/layouts/vamcSystem/ManageYourHealthLinks'
+import { ManageYourHealthLinks } from './ManageYourHealthLinks'
 import { LovellSwitcher } from '@/templates/components/lovellSwitcher'
 import { StoryTeaser } from './StoryTeaser'
-import FacilitySocialLinks from '../healthCareLocalFacility/FacilitySocialLinks'
+import FacilitySocialLinks from '@/templates/layouts/healthCareLocalFacility/FacilitySocialLinks'
 // import { ListOfLinkTeasers } from '@/templates/components/listOfLinkTeasers'
 import { EventTeaser } from '@/templates/components/eventTeaser'
 // import { SocialLinks } from '@/templates/common/socialLinks'
